error 24/10/2024

+++ restaurant-android/app/src/main/java/com/ktpm1/restaurant/models/Order.java
@@ -18,7 +18,7 @@ public class Order {
     User user;
     Table table;
     LocalDateTime orderTime;
-    long totalPrice;
+    int totalPrice;
     OrderStatus status;
     Set<OrderItem> orderItems = new HashSet<>();
     boolean paid = false;
