+++ restaurant-android/app/src/main/java/com/ktpm1/restaurant/models/OrderItem.java
@@ -15,6 +15,6 @@ public class OrderItem {
     Order order;
     Food food;
     int quantity;
-    long itemPrice;
+    int itemPrice;
 }
